backend/requirements.txt
Django==4.2.6
djangorestframework==3.16.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.7
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
pytest-django==4.8.0
factory-boy==3.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
mixer==7.2.2